Doc changes would be located there if we intended to contributed them upstream. Anything on this branch is specific to Teradata's version and won't go upstream.

Keeping this as a merge commit makes this easier to determine which is which.

Merge branch 'phase2-on-0.127' into teradata-docs

Conflicts:
	presto-docs/src/main/sphinx/admin.rst
	presto-docs/src/main/sphinx/connector/hive.rst
	presto-docs/src/main/sphinx/develop.rst
	presto-docs/src/main/sphinx/develop/example-http.rst
	presto-docs/src/main/sphinx/develop/functions.rst
	presto-docs/src/main/sphinx/develop/spi-overview.rst
	presto-docs/src/main/sphinx/develop/types.rst
	presto-docs/src/main/sphinx/functions/decimal.rst
	presto-docs/src/main/sphinx/installation.rst
	presto-docs/src/main/sphinx/language/types.rst

Also removed:
     src/main/sphinx/admin/kerberos.rst
     src/main/sphinx/installation/deployment-on-yarn.rst

+++ presto-main/src/main/java/com/facebook/presto/connector/system/jdbc/ColumnJdbcTable.java
@@ -128,19 +128,19 @@ public class ColumnJdbcTable
                     tableName.getSchemaName(),
                     tableName.getTableName(),
                     column.getName(),
-                    jdbcDataType(column.getType()),
+                    (long) jdbcDataType(column.getType()),
                     column.getType().getDisplayName(),
-                    0,
-                    0,
+                    0L,
+                    0L,
                     decimalDigits(column.getType()),
-                    10,
+                    10L,
                     DatabaseMetaData.columnNullableUnknown,
                     column.getComment(),
                     null,
                     null,
                     null,
-                    0,
-                    ordinalPosition,
+                    0L,
+                    (long) ordinalPosition,
                     "",
                     null,
                     null,
@@ -190,10 +190,10 @@ public class ColumnJdbcTable
         return Types.JAVA_OBJECT;
     }
 
-    private static Integer decimalDigits(Type type)
+    private static Long decimalDigits(Type type)
     {
         if (type.equals(BIGINT)) {
-            return 19;
+            return 19L;
         }
         return null;
     }
